+++ src/main/java/chess/controller/ChessController.java
@@ -1,8 +1,8 @@
 package chess.controller;
 
-import chess.model.ChessBoard;
-import chess.model.ChessBoardInitializer;
-import chess.model.ChessPosition;
+import chess.model.board.ChessBoard;
+import chess.model.board.ChessBoardInitializer;
+import chess.model.position.ChessPosition;
 import chess.view.GameArguments;
 import chess.view.GameCommand;
 import chess.view.InputView;
