+++ src/main/java/chess/model/piece/Rook.java
@@ -1,28 +1,30 @@
-package chess.model;
+package chess.model.piece;
+
+import chess.model.position.ChessPosition;
+import chess.model.position.Distance;
 
 import java.util.List;
 
-public class King extends Piece {
-    private static final int DISPLACEMENT = 1;
+public class Rook extends Piece {
 
-    public King(Side side) {
+    public Rook(Side side) {
         super(side);
     }
 
     @Override
     public String getText() {
         if (side.isWhite()) {
-            return "k";
+            return "r";
         }
-        return "K";
+        return "R";
     }
 
     @Override
     public List<ChessPosition> findPath(ChessPosition source, ChessPosition target, Piece targetPiece) {
         checkValidTargetPiece(targetPiece);
         Distance distance = target.calculateDistance(source);
-        if (distance.hasSame(DISPLACEMENT)) {
-            return List.of(target);
+        if (distance.isCrossMovement()) {
+            return distance.findPath(source);
         }
         return List.of();
     }
