+++ src/main/java/chess/model/piece/Side.java
@@ -1,4 +1,4 @@
-package chess.model;
+package chess.model.piece;
 
 public enum Side {
     BLACK,
