+++ src/main/java/chess/model/position/Rank.java
@@ -1,4 +1,6 @@
-package chess.model;
+package chess.model.position;
+
+import chess.model.piece.Side;
 
 import java.util.Arrays;
