+++ src/test/java/chess/model/board/ChessBoardInitializerTest.java
@@ -1,10 +1,13 @@
-package chess.model;
+package chess.model.board;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.util.Map;
+import chess.model.piece.Piece;
+import chess.model.position.ChessPosition;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import java.util.Map;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 
 class ChessBoardInitializerTest {
