+++ src/test/java/chess/model/position/ChessPositionTest.java
@@ -1,7 +1,11 @@
-package chess.model;
+package chess.model.position;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import chess.model.position.ChessPosition;
+import chess.model.position.Distance;
+import chess.model.position.File;
+import chess.model.position.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
