+++ src/test/java/chess/model/position/DistanceTest.java
@@ -1,4 +1,4 @@
-package chess.model;
+package chess.model.position;
 
 import static org.assertj.core.api.Assertions.assertThat;
