+++ src/test/java/chess/model/position/RankTest.java
@@ -1,7 +1,9 @@
-package chess.model;
+package chess.model.position;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import chess.model.piece.Side;
+import chess.model.position.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
